requirements.txt
# PyTorch (CUDA recommended for SAM2 and Grounding DINO)
torch>=2.0
torchvision>=0.15

# Image Processing
numpy<2
opencv-python>=4.8
Pillow>=10.0

# Configuration
PyYAML>=6.0

# Transformers & Grounding DINO
transformers>=4.40
sentencepiece>=0.1.99
tokenizers>=0.13.3

# SAM2 Dependencies
timm>=0.9.0
scipy>=1.11.2


# Optional: Progress bars
tqdm>=4.66.1

# yolo training
ultralytics

mcap-ros2-support

torch --index-url https://download.pytorch.org/whl/cu124
